Style: move some necessary things to constants
