+++ src/main/java/frc/lib/generic/pigeon/Pigeon.java
@@ -1,6 +1,6 @@
 package frc.lib.generic.pigeon;
 
-import frc.lib.generic.advantagekit.HardwareManager;
+import frc.lib.generic.HardwareManager;
 import frc.lib.generic.advantagekit.LoggableHardware;
 import org.littletonrobotics.junction.Logger;
