+++ src/main/java/frc/robot/GlobalConstants.java
@@ -14,7 +14,9 @@ public class GlobalConstants {
     public static final Lock FASTER_THREAD_LOCK = new ReentrantLock();
 
     public static final boolean IS_TUNING_MODE = true;
-    public static final Mode CURRENT_MODE = Mode.REAL;
+
+    public static final boolean SHOULD_WRITE_LOGS = false;  //for when the roborio doesn't have enough space...
+    public static final Mode CURRENT_MODE = Mode.SIMULATION;
 
     public enum Mode {
         REAL, SIMULATION, REPLAY
